+++ src/record.h
@@ -12,11 +12,11 @@ class Record
 {
 	public:
 		Record();
-		Record(int acc, std::string nm, int blnc);
+		Record(int acc, std::string nm, double blnc);
 		Record(std::string  record);
 		int account;
 		std::string name;
-		int balance;
+		double balance;
 		int is_locked;
 		pthread_cond_t in_use;
 };
