+++ src/transaction.h
@@ -12,13 +12,13 @@ class Transaction
 {
 	public:
 		Transaction();
-		Transaction(int tm, int acc, std::string tp, int amt);
+		Transaction(int tm, int acc, std::string tp, double amt);
 		Transaction(std::string transaction);
 		int is_valid();
 		int time;
 		int account;
 		std::string type;
-		int amount;
+		double amount;
 };
 
 #endif
